(6/21) SECOND RELEASE, SOME BUG FIXED and SOME QUESTION

+++ include/linux_parser.h
@@ -55,7 +55,7 @@ enum ProcStates {
 std::vector<std::string> CpuUtilization();
 long Jiffies();
 long ActiveJiffies();
-long ActiveJiffies(int pid);
+float ActiveJiffies(int pid);
 long IdleJiffies();
